requirement.txt

streamlit
pandas
psycopg2-binary
python-dotenv
dbt-core
requests
pyyaml
